requirements.txt
asgiref==3.7.2
certifi==2023.5.7
distlib==0.3.6
dj-database-url==2.0.0
Django==4.2.3
django-cors-headers==4.2.0
django-heroku==0.3.1
djangorestframework==3.14.0
filelock==3.12.2
gunicorn==20.1.0
pipenv==2023.7.4
platformdirs==3.8.1
psycopg2-binary==2.9.6
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
virtualenv==20.23.1
virtualenv-clone==0.5.7
whitenoise==6.5.0
